refactor(guid): 将 constants.ts 移至 utils 子目录
