+++ src/renderer/pages/guid/components/AssistantSelectionArea.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { resolveExtensionAssetUrl } from '@/renderer/utils/platform';
 import type { AcpBackendConfig, AvailableAgent } from '../types';
-import { CUSTOM_AVATAR_IMAGE_MAP } from '../constants';
+import { CUSTOM_AVATAR_IMAGE_MAP } from '../utils/constants';
 
 type AssistantSelectionAreaProps = {
   customAgents: AcpBackendConfig[];
