+++ src/renderer/pages/guid/hooks/useGuidMention.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { getAgentLogo } from '@/renderer/utils/agentLogo';
-import { CUSTOM_AVATAR_IMAGE_MAP } from '../constants';
+import { CUSTOM_AVATAR_IMAGE_MAP } from '../utils/constants';
 import type { AvailableAgent, MentionOption } from '../types';
 
 export type GuidMentionResult = {
